+++ docs/changelog.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+## (in development) 2021.07.30
+
+- Force system transport ssh connections to allocate a tty (-tt); fixes issue that would prevent system transport 
+  from sending any command > 1024 chars.
+
+
 # 2021.01.30
 
 - Big overhaul in line with the scrapli core overhaul... mostly this was about reconciliation of the channel and